# faro_backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: Faro MiniGPT API
  version: 1.0.0
  description: GPT Actions endpoints for Faro — city recommendations, comparisons, grants, and action plans for underrepresented entrepreneurs.
servers:
  - url: https://minigpt.farosmart.com/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      description: Returns ok if the API is live.
      security: []
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /chat:
    post:
      operationId: postChat
      summary: Chat with Faro
      description: Send a message to Faro and get a conversational response about city relocation and entrepreneurship.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              additionalProperties: false
              properties:
                message:
                  type: string
                history:
                  type: array
                  items:
                    type: object
                    additionalProperties: false
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
                    required: [role, content]
      responses:
        "200":
          description: Chat reply
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /recommend:
    post:
      operationId: postRecommend
      summary: Recommend U.S. cities
      description: Returns a ranked list of U.S. cities based on entrepreneur context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                industry:
                  type: string
                stage:
                  type: string
                budget:
                  type: number
                timeline:
                  type: string
                priorities:
                  type: array
                  items:
                    type: string
                targetCustomer:
                  type: string
                shortlistCities:
                  type: array
                  items:
                    type: string
                notes:
                  type: string
      responses:
        "200":
          description: City recommendations
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  cities:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        name:
                          type: string
                        score:
                          type: integer
                          minimum: 0
                          maximum: 100
                        why:
                          type: string
                      required: [name, score, why]
                  followUps:
                    type: array
                    items:
                      type: string
                required: [cities]
        "401":
          description: Unauthorized
        "502":
          description: AI response invalid
  /compare-cities:
    post:
      operationId: postCompareCities
      summary: Compare two U.S. cities
      description: Compares two cities with pros, cons, scores, and a verdict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [city1, city2]
              additionalProperties: false
              properties:
                city1:
                  type: string
                city2:
                  type: string
                industry:
                  type: string
                budget:
                  type: string
                priorities:
                  type: string
      responses:
        "200":
          description: City comparison
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  city1:
                    type: string
                  city2:
                    type: string
                  summary:
                    type: string
                  verdict:
                    type: string
                  scores:
                    type: object
                    properties:
                      city1:
                        type: integer
                      city2:
                        type: integer
                  pros:
                    type: object
                    properties:
                      city1:
                        type: array
                        items:
                          type: string
                      city2:
                        type: array
                        items:
                          type: string
                  cons:
                    type: object
                    properties:
                      city1:
                        type: array
                        items:
                          type: string
                      city2:
                        type: array
                        items:
                          type: string
                  nextSteps:
                    type: array
                    items:
                      type: string
        "400":
          description: Missing city1 or city2
        "401":
          description: Unauthorized
        "502":
          description: AI response invalid
  /grants/search:
    post:
      operationId: postGrantsSearch
      summary: Search grants and funding
      description: Returns relevant grants and funding programs for a given city, industry, and stage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                city:
                  type: string
                industry:
                  type: string
                stage:
                  type: string
                keywords:
                  type: string
      responses:
        "200":
          description: Grants results
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  grants:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        name:
                          type: string
                        amount:
                          type: string
                        deadline:
                          type: string
                          nullable: true
                        eligibility:
                          type: string
                        city:
                          type: string
                          nullable: true
                        industry:
                          type: string
                          nullable: true
                        link:
                          type: string
                          nullable: true
                      required: [name, amount, eligibility]
                  disclaimer:
                    type: string
                required: [grants, disclaimer]
        "401":
          description: Unauthorized
        "502":
          description: AI response invalid
  /action-plan:
    post:
      operationId: postActionPlan
      summary: Generate a city action plan
      description: Generates a step-by-step relocation and business launch action plan for a specific city and industry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [city, industry]
              additionalProperties: false
              properties:
                city:
                  type: string
                industry:
                  type: string
                stage:
                  type: string
                budget:
                  type: number
                timeline:
                  type: string
                priorities:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Action plan
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  city:
                    type: string
                  industry:
                    type: string
                  summary:
                    type: string
                  steps:
                    type: array
                    items:
                      type: object
                      properties:
                        order:
                          type: integer
                        action:
                          type: string
                        timeline:
                          type: string
                        resources:
                          type: string
                          nullable: true
                  milestones:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: integer
                        goal:
                          type: string
                  warnings:
                    type: array
                    items:
                      type: string
                    nullable: true
                required: [city, industry, summary, steps, milestones]
        "400":
          description: Missing city or industry
        "401":
          description: Unauthorized
        "502":
          description: AI response invalid